pyproject.toml
[project]
name = "news-digest-mcp-bot"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "Ittichai Boonyarakthunya",email = "i.boonyarakthunya@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "huggingface (>=0.0.1,<0.0.2)",
    "pinecone-client (>=6.0.0,<7.0.0)",
    "dvc (>=3.59.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

